+++ Makefile
@@ -1,7 +1,7 @@
 COMPILER = gcc
 TARGET = kvmm
 CFLAGS = -pthread
-OBJECTS = main.o util.o blk.o io.o
+OBJECTS = main.o util.o blk.o io.o lapic.o interrupt.o
 
 all: 
 	make $(TARGET)
