+++ blk.h
@@ -1,6 +1,10 @@
 #include "type.h"
 #include "vcpu.h"
 #include "io.h"
+#include "interrupt.h"
+#include "lapic.h"
+
+#define IMGE_SIZE 5120000
 
 struct blk {
     u8 *data;
@@ -17,3 +21,6 @@ struct blk {
 
 void emulate_diskr(struct vcpu *vcpu, struct blk *blk);
 void emulate_diskw(struct vcpu *vcpu, struct blk *blk, struct io io);
+void update_blk_index(struct blk *blk);
+void emulate_disk_portw(struct vcpu *vcpu, struct blk *blk, struct io io);
+void emulate_disk_portr(struct vcpu *vcpu, struct blk *blk);
