+++ main.c
@@ -21,7 +21,7 @@
 #define ALIGNMENT_SIZE 0x1000
 #define DEFAULT_FLAGS 0x0000000000000002ULL
 #define GUEST_BINARY_SIZE (4096 * 128)
-#define IMGE_SIZE 5120000
+// #define IMGE_SIZE 5120000
 #define FS_IMAGE_SIZE (500 * 1024)
 #define LAPIC_BASE 0xfee00000
 #define IOAPIC_BASE 0xfec00000
@@ -109,19 +109,9 @@ struct ioapic {
     union redirtb_entry redirtb[24];
 };
 
-struct irr_queue {
-    int buff[1000];
-    int last;
-};
-
 struct interrupt_buffer {
     int head, end, count, max;
     int *buff;
-};
-
-struct lapic {
-    u32 regs[1024];
-    struct irr_queue *irr;
 };
 
 int outfd = 0;
@@ -133,23 +123,23 @@ struct blk *blk;
 struct uart *uart;
 int irr_count = 0;
 
-void enq_irr(struct irr_queue *irr, int value) {
-    irr->buff[irr->last] = value;
-    irr->last++;
-    irr_count++;
-}
+// void enq_irr(struct irr_queue *irr, int value) {
+//     irr->buff[irr->last] = value;
+//     irr->last++;
+//     irr_count++;
+// }
 
-int deq_irr(struct irr_queue *irr) {
-   int out = irr->buff[0];
-   for (int i = 0; i <= irr->last; i++) {
-       irr->buff[i] = irr->buff[i+1];
-   }
-   irr->last--;
+// int deq_irr(struct irr_queue *irr) {
+//    int out = irr->buff[0];
+//    for (int i = 0; i <= irr->last; i++) {
+//        irr->buff[i] = irr->buff[i+1];
+//    }
+//    irr->last--;
 
-    if (irr_count > 0) {
-        irr_count--;
-    }
-}   
+//     if (irr_count > 0) {
+//         irr_count--;
+//     }
+// }   
 
 // void error(char *message) {
 //     perror(message);
@@ -388,68 +378,68 @@ void create_output_file() {
 //     }
 // }
 
-void update_blk_index(struct blk *blk) {
-    u32 index = 0 | blk->lba_low_reg | (blk->lba_middle_reg << 8) | (blk->lba_high_reg << 16);
-    blk->index = index * 512;
-    if (blk->drive_head_reg == 0xf0) {
-        blk->index += IMGE_SIZE;
-    }
-}
+// void update_blk_index(struct blk *blk) {
+//     u32 index = 0 | blk->lba_low_reg | (blk->lba_middle_reg << 8) | (blk->lba_high_reg << 16);
+//     blk->index = index * 512;
+//     if (blk->drive_head_reg == 0xf0) {
+//         blk->index += IMGE_SIZE;
+//     }
+// }
 
-void emulate_disk_portw(struct vcpu *vcpu, 
-                        struct blk *blk, struct io io) {
-    u8 val1;
-    u16 val2;
+// void emulate_disk_portw(struct vcpu *vcpu, 
+//                         struct blk *blk, struct io io) {
+//     u8 val1;
+//     u16 val2;
 
-    val1 = *(u8*)((u8*)vcpu->kvm_run + io.data_offset);
+//     val1 = *(u8*)((u8*)vcpu->kvm_run + io.data_offset);
     
-    switch (io.port) {
-    case 0x1F0:
-        emulate_diskw(vcpu, blk, io);
-        break;
-    case 0x1F2:
-        blk->sec_count_reg = val1;
-        break;
-    case 0x1F3:
-        blk->lba_low_reg = val1;
-        break;            
-    case 0x1F4:
-        blk->lba_middle_reg = val1;
-        break;
-    case 0x1F5:
-        blk->lba_high_reg = val1;
-    case 0x1F6:
-        blk->drive_head_reg = val1;
-        break;
-    case 0x3F6:
-        blk->dev_conotrl_regs = val1;
-    default:
-        break;
-    }
+//     switch (io.port) {
+//     case 0x1F0:
+//         emulate_diskw(vcpu, blk, io);
+//         break;
+//     case 0x1F2:
+//         blk->sec_count_reg = val1;
+//         break;
+//     case 0x1F3:
+//         blk->lba_low_reg = val1;
+//         break;            
+//     case 0x1F4:
+//         blk->lba_middle_reg = val1;
+//         break;
+//     case 0x1F5:
+//         blk->lba_high_reg = val1;
+//     case 0x1F6:
+//         blk->drive_head_reg = val1;
+//         break;
+//     case 0x3F6:
+//         blk->dev_conotrl_regs = val1;
+//     default:
+//         break;
+//     }
 
-    update_blk_index(blk);
+//     update_blk_index(blk);
 
-    if ((io.port == 0x1F0 || io.port == 0x1F7) && blk->dev_conotrl_regs == 0) {
-        enq_irr(lapic->irr,IRQ_BASE+14);
-        vcpu->kvm_run->request_interrupt_window = 1;
-    }
-}
+//     if ((io.port == 0x1F0 || io.port == 0x1F7) && blk->dev_conotrl_regs == 0) {
+//         enq_irr(lapic->irr,IRQ_BASE+14);
+//         vcpu->kvm_run->request_interrupt_window = 1;
+//     }
+// }
 
-void emulate_disk_portr(struct vcpu *vcpu,
-                        struct blk *blk) {
-    u32 data = 0;
+// void emulate_disk_portr(struct vcpu *vcpu,
+//                         struct blk *blk) {
+//     u32 data = 0;
 
-    switch (vcpu->kvm_run->io.port) {
-    case 0x1F0:
-        emulate_diskr(vcpu, blk);
-        break;
-    case 0x1F7:
-         *(unsigned char*)((unsigned char*)vcpu->kvm_run + vcpu->kvm_run->io.data_offset) = blk->status_command_reg; 
-        break;
-    default:
-        break;
-    }
-}
+//     switch (vcpu->kvm_run->io.port) {
+//     case 0x1F0:
+//         emulate_diskr(vcpu, blk);
+//         break;
+//     case 0x1F7:
+//          *(unsigned char*)((unsigned char*)vcpu->kvm_run + vcpu->kvm_run->io.data_offset) = blk->status_command_reg; 
+//         break;
+//     default:
+//         break;
+//     }
+// }
 
 int write_f = 0;
